+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayerOptions.java
@@ -4,8 +4,26 @@
 
 package io.flutter.plugins.videoplayer;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 public class VideoPlayerOptions {
   public boolean mixWithOthers;
   public int maxLoadRetries = 5;
   public int maxPlayerRecoveryAttempts = 3;
+
+  /**
+   * The duration of the back buffer in milliseconds, used to configure ExoPlayer's load control.
+   */
+  @Nullable public Long backBufferDurationMs;
+
+  public VideoPlayerOptions() {}
+
+  /** Copy constructor to ensure all options are reliably copied. */
+  public VideoPlayerOptions(@NonNull VideoPlayerOptions other) {
+    this.mixWithOthers = other.mixWithOthers;
+    this.maxLoadRetries = other.maxLoadRetries;
+    this.maxPlayerRecoveryAttempts = other.maxPlayerRecoveryAttempts;
+    this.backBufferDurationMs = other.backBufferDurationMs;
+  }
 }