+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation_objc/FVPVideoPlayer.m
@@ -910,6 +910,11 @@ NS_INLINE CGFloat radiansToDegrees(CGFloat radians) {
   NSLog(@"PiP failed to start: %@", error.localizedDescription);
 }
 
+- (void)setPreventsDisplaySleepDuringVideoPlayback:(BOOL)preventsDisplaySleepDuringVideoPlayback
+                                             error:(FlutterError *_Nullable *_Nonnull)error {
+  self.player.preventsDisplaySleepDuringVideoPlayback = preventsDisplaySleepDuringVideoPlayback;
+}
+
 #pragma mark - Private
 
 - (int64_t)duration {